Fix the desynchronization of results between instant combat and auto/quick combat

+++ src/fheroes2/battle/battle_arena.cpp
@@ -515,8 +515,8 @@ void Battle::Arena::UnitTurn( const Units & orderHistory )
             }
         }
 
-        const uint32_t newSeed = std::accumulate( actions.cbegin(), actions.cend(), static_cast<uint32_t>( _randomGenerator.getStream() ),
-                                                  []( const uint32_t seed, const Command & cmd ) { return cmd.updateSeed( seed ); } );
+        const uint64_t newSeed = std::accumulate( actions.cbegin(), actions.cend(), _randomGenerator.getStream(),
+                                                  []( const uint64_t seed, const Command & cmd ) { return cmd.updateSeed( seed ); } );
         _randomGenerator.setStream( newSeed );
 
         while ( !actions.empty() ) {
